Moved robotmap to new constants class
