+++ src/main/java/frc/robot/Constants.java
@@ -16,4 +16,48 @@ package frc.robot;
  * constants are needed, to reduce verbosity.
  */
 public final class Constants {
+    public static class RobotMap {
+        public static class CAN {
+            public static final int LEFT_FRONT_MOTOR = 37;
+            public static final int LEFT_BACK_MOTOR = 46;
+            public static final int RIGHT_FRONT_MOTOR = 34;
+            public static final int RIGHT_BACK_MOTOR = 35;
+    
+            public static final int SHOOTER_1 = 0;
+            public static final int SHOOTER_2 = 0;
+        }
+    
+        public static class DIO {
+        
+        }
+    
+        public static class PCM {
+    
+        }
+    
+        public static class XBOX {
+            // Buttons
+            public static final int BUTTON_A = 1;
+            public static final int BUTTON_B = 2;
+            public static final int BUTTON_X = 3;
+            public static final int BUTTON_Y = 4;
+            public static final int BUMPER_LEFT = 5;
+            public static final int BUMPER_RIGHT = 6;
+            public static final int BUTTON_BACK = 7;
+            public static final int BUTTON_START = 8;
+            public static final int STICK_LEFT = 9;
+            public static final int STICK_RIGHT = 10;
+    
+            // Axes
+            public static final int STICK_L_X_AXIS = 0;
+            public static final int STICK_L_Y_AXIS = 1;
+            public static final int STICK_R_X_AXIS = 4;
+            public static final int STICK_R_Y_AXIS = 5;
+            public static final int TRIGGER_L_AXIS = 2;
+            public static final int TRIGGER_R_AXIS = 3;
+        }
+    }
+    public static class DriveConstants {
+        
+    }
 }
