+++ src/main/java/frc/robot/subsystems/Drivetrain.java
@@ -19,7 +19,7 @@ import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
-import frc.robot.RobotMap;
+import frc.robot.Constants.RobotMap;
 
 public class Drivetrain extends SubsystemBase {
   private DifferentialDrive robotDrive;
